fix: get rid of transformer package
