+++ apps/loadtest/package.json
@@ -8,7 +8,6 @@
   },
   "dependencies": {
     "@mp/network": "workspace:*",
-    "@mp/transformer": "workspace:*",
     "@mp/server": "workspace:*",
     "@mp/logger": "workspace:*",
     "@mp/data": "workspace:*",
