+++ apps/server/package.json
@@ -25,7 +25,6 @@
     "@mp/state": "workspace:*",
     "@mp/tiled-loader": "workspace:*",
     "@mp/time": "workspace:*",
-    "@mp/transformer": "workspace:*",
     "@trpc/server": "11.0.0-rc.593",
     "cors": "2.8.5",
     "dotenv-flow": "4.1.0",
@@ -33,6 +32,7 @@
     "express": "4.19.2",
     "postgres": "^3.4.4",
     "rate-limiter-flexible": "^5.0.3",
+    "superjson": "2.2.1",
     "yargs": "17.7.2"
   },
   "devDependencies": {
