+++ apps/server/src/context.ts
@@ -1,7 +1,7 @@
 import type { Branded } from "@mp/data";
 import type { AuthClient } from "@mp/auth/server";
 import type { Logger } from "@mp/logger";
-import type { StateAccess } from "@mp/transformer";
+import type { StateAccess } from "@mp/network/server";
 import type { WorldState } from "./modules/world/schema";
 import type { ClientRegistry } from "./modules/world/ClientRegistry";
