+++ apps/server/src/transformer.ts
@@ -1,1 +1,1 @@
-export { SuperJSON as transformer } from "@mp/transformer";
+export { default as transformer } from "superjson";
