+++ packages/network/package.json
@@ -25,11 +25,13 @@
   },
   "dependencies": {
     "@mp/state": "workspace:*",
+    "@automerge/automerge": "2.2.8",
     "socket.io": "4.7.5",
     "socket.io-client": "4.7.5"
   },
   "devDependencies": {
     "@mp/tsconfig": "workspace:*",
+    "type-fest": "4.26.1",
     "vitest": "2.0.3"
   }
 }
