--- packages/transformer/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@mp/transformer",
-  "version": "0.0.0",
-  "private": true,
-  "exports": {
-    ".": {
-      "import": "./src/index.ts",
-      "require": "./src/index.ts"
-    }
-  },
-  "scripts": {
-    "test": "vitest run",
-    "test:watch": "vitest"
-  },
-  "dependencies": {
-    "cbor2": "1.7.0",
-    "pako": "2.1.0",
-    "superjson": "2.2.1",
-    "@automerge/automerge": "2.2.8"
-  },
-  "devDependencies": {
-    "@mp/tsconfig": "workspace:*",
-    "@types/pako": "2.0.3",
-    "type-fest": "4.26.1",
-    "vitest": "2.0.3"
-  }
-}
