--- packages/transformer/src/index.ts
@@ -1,6 +1,0 @@
-export * from "./transformer";
-export * from "./json";
-export * from "./cbor";
-export * from "./decoders";
-export * from "./decompressors";
-export * from "./crdt";
