--- packages/transformer/test/transformer.test.ts
@@ -1,5 +1,0 @@
-import { it, expect } from "vitest";
-
-it("can run test inside the transformer package", () => {
-  expect(true).toEqual(true);
-});
