--- packages/transformer/test/tsconfig.json
@@ -1,1 +1,0 @@
-{ "extends": ["@mp/tsconfig/node.json"] }
